Update release utility files

+++ fix-jlcpcb-files.sh
@@ -7,13 +7,17 @@ PROJ="SH-RPi"
 DESTDIR="assembly"
 BOMFILE="${PROJ}_bom_jlc.csv"
 CPLFILE="${PROJ}_cpl_jlc.csv"
+BOM_SRC_FILE="${PROJ}.xml"
 
-# fix the BOM and CPL files
-jlc-kicad-tools -n $PROJ -d cpl_rotations_db.csv .
+# if we're ordering PCBs only, ignore the BOM and CPL parts
+if [ -f $BOM_SRC_FILE ] ; then
+    # fix the BOM and CPL files
+    jlc-kicad-tools -n $PROJ -d cpl_rotations_db.csv .
 
-# move the BOM and CPL files
-mkdir -p $DESTDIR
-mv $BOMFILE $CPLFILE $DESTDIR
+    # move the BOM and CPL files
+    mkdir -p $DESTDIR
+    mv $BOMFILE $CPLFILE $DESTDIR
+fi
 
 # zip the DESTDIR directory
 zip -FSr $DESTDIR.zip $DESTDIR/
